Fix bad build.gradle configs and refactored directory
