+++ README.md
@@ -1,3 +1,4 @@
 ### Tips
 1. Creating a jar `./gradlew clean shadowJar`. Clean is optional 
 2. `java -jar {filename}.jar` - Runs a jar file
+3.
