+++ src/main/java/baron/Main.java
@@ -1,0 +1,12 @@
+package baron;
+
+import javafx.application.Application;
+
+/**
+ * Entry point file to the chatbot.
+ */
+public class Main {
+    public static void main(String[] args) {
+        Application.launch(Launcher.class, args);
+    }
+}
